+++ src/lib/stress.js
@@ -1,0 +1,81 @@
+'use strict';
+
+module.exports = class Stress {
+
+	static get maxStress() {
+		return 100;
+	}
+
+	/**
+	 * Stress Simulator (Isn't the real life?)
+	 * @param {Object} options
+	 * @param {Number} options.stressMultiplier The stress multiplier factor (the more sh*t happens, the more stress you get, compadre)
+	 * @param {Number} options.copingTime The time (seconds) to wait until the stress returns to <= 1 (the amount of time you need to cry)
+	 * @example
+	 * // Increase stress proportionally at 1.25x and cry for 2 minutes to feel good again
+	 * new Stress({ stressMultiplier: 1.25, copingTime: 120 });
+	 */
+	constructor({ stressMultiplier, copingTime }) {
+
+		this.copingCooldownFactor = this.constructor.maxStress / copingTime;
+		this.copingMechanism = {};
+
+		this.stressMultiplier = stressMultiplier;
+		this.stressLevel = 1;
+	}
+
+	/**
+	 * Applies pressure to you, like every single thing that happens in real life. Yeah let's cry together (Imagine crying btw)
+	 * @returns {Boolean}
+	 * - `true` If despite your stress level you can continue faking that smile
+	 * - `false` If you're f*cked up :(
+	 */
+	pressure() {
+
+		if(this.isSuffering)
+			return false;
+
+		this.increaseStress();
+
+		return !this.isSuffering;
+	}
+
+	/**
+	 * HAHAHAHA YES >:D
+	 * - Uh, yeah this increases the stress level using the multipler and that, you know...
+	 */
+	increaseStress() {
+
+		this.stressLevel *= this.stressMultiplier;
+
+		if(this.stressLevel >= this.constructor.maxStress)
+			this.isSuffering = true;
+
+		this.cope();
+	}
+
+	/**
+	 * Yeah! Bottle your problems and feelings, how long can you last? *laughs in crying*
+	 * - Also this basically generates the coping mechanism which reduces the stress based on the specified coping time. **WOAH I'M SO NEEEEERDDD OWO**
+	 * - NO, no, I'm not good at math. Yes, this has nothing to do with here, but I wanted to say it
+	 */
+	cope() {
+
+		if(this.copingMechanism._destroyed === false)
+			return;
+
+		this.copingMechanism = setInterval(() => {
+
+			this.stressLevel -= this.copingCooldownFactor;
+
+			if(this.stressLevel > 1)
+				return;
+
+			this.isSuffering = false;
+			this.stressLevel = 1;
+
+			clearInterval(this.copingMechanism);
+
+		}, 1000);
+	}
+};